// src/app/core/models/Haghighi/HaghighiAddress.ts
export interface HaghighiAddress {
    Id?: number;
    ShakhsHaghighiId?: number;
    CityId?: any;
    CityName?: any;
    ProvinceId?: number;
    ProvinceCode?: any;
    ProvinceName?: any;
    PostalCode?: string;
    Address?: string;
    // TelephoneNumber?: string;
    ShomarehTelephoneSabet?: string;
}